requirements.txt

Django==3.2.9

Pillow==9.0.1

django-modeltranslation